+++ substrate/frame/support/test/tests/genesisconfig.rs
@@ -63,7 +63,7 @@ pub mod pallet {
 	}
 }
 
-pub type BlockNumber = u32;
+pub type BlockNumber = u64;
 pub type Signature = sr25519::Signature;
 pub type AccountId = <Signature as Verify>::Signer;
 pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
