+++ substrate/frame/support/test/tests/pallet_ui/pass/no_std_genesis_config.rs
@@ -20,7 +20,7 @@ use sp_core::sr25519;
 use sp_runtime::{generic, traits::BlakeTwo256};
 
 pub type Signature = sr25519::Signature;
-pub type BlockNumber = u32;
+pub type BlockNumber = u64;
 pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
 pub type Block = generic::Block<Header, UncheckedExtrinsic>;
 pub type UncheckedExtrinsic = generic::UncheckedExtrinsic<u32, RuntimeCall, Signature, ()>;
